Fixed the hang bug when skipping to next plate early
The code for NextPlateOverride was not marking the current well
(status->started) as skipped or anything, leading to a hang up
with the subsequent program function. Altered the code to change
both 'started' and 'uncomplete' in the current plate as 'skipped'
and program flow is restored. Now, one needs to click 'next' to
move to the next step, but I want to fix that.